API

Vulnerability curve

SoilHydraulics.relative_hydraulic_conductance — Method

relative_hydraulic_conductance(vc::ComplexVC{FT}, p_25::FT) where {FT<:AbstractFloat}

Return the hydraulic conductance ralative to maximum at reference temperature, given

vc ClimaCache.ComplexVC type vulnerability curve

p_25 Equivalent xylem water pressu

Examples

vc = ComplexVC{Float64}([0.3,0.3,0.4], [LogisticVC{FT}(2,2), PowerVC{FT}(2,2), WeibullVC{FT}(2,2)]);
kr = relative_hydraulic_conductance(vs4, FT(-2.0));

SoilHydraulics.relative_hydraulic_conductance — Method

relative_hydraulic_conductance(vc::LogisticVC{FT}, p_25::FT) where {FT<:AbstractFloat}

Return the hydraulic conductance ralative to maximum at reference temperature, given

vc ClimaCache.LogisticVC type vulnerability curve

p_25 Equivalent xylem water pressure at 298.15 K in [MPa] (surface tension correction made)

Examples

vc = LogisticVC{Float64}(2, 2);
kr = relative_hydraulic_conductance(vc, -2.0);

SoilHydraulics.relative_hydraulic_conductance — Method

relative_hydraulic_conductance(vc::PowerVC{FT}, p_25::FT) where {FT<:AbstractFloat}

Return the hydraulic conductance ralative to maximum at reference temperature, given

vc ClimaCache.PowerVC type vulnerability curve

p_25 Equivalent xylem water pressure at 298.15 K in [MPa] (surface tension correction made)

Examples

vc = PowerVC{Float64}(2, 2);
kr = relative_hydraulic_conductance(vc, -2.0);

SoilHydraulics.relative_hydraulic_conductance — Method

relative_hydraulic_conductance(vc::WeibullVC{FT}, p_25::FT) where {FT<:AbstractFloat}

Return the hydraulic conductance ralative to maximum at reference temperature, given

vc ClimaCache.WeibullVC type vulnerability curve

p_25 Equivalent xylem water pressure at 298.15 K in [MPa] (surface tension correction made)

Examples

vc = WeibullVC{Float64}(2, 2);
kr = relative_hydraulic_conductance(vc, -2.0);

PlantHydraulics.critical_pressure — Function

This function returns the critical xylem water pressure (at 25 °C) that triggers 99.triggers a given amount of loss of hydraulic conductance. The supported methods are

critical_pressure(vc)
critical_pressure(vc, kr)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/critical_pressure.jl:36.

critical_pressure(vc)
critical_pressure(vc, kr)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/critical_pressure.jl:64.

critical_pressure(vc)
critical_pressure(vc, kr)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/critical_pressure.jl:87.

critical_pressure(vc)
critical_pressure(vc, kr)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/critical_pressure.jl:110.

PlantHydraulics.critical_pressure — Method

critical_pressure(vc::ComplexVC{FT}, kr::FT = FT(0.001)) where {FT<:AbstractFloat}

Return the critical xylem water pressure at 25 °C that triggers a given amount of loss of conductance, given

vc ClimaCache.ComplexVC type struct

kr Reference conductance, default is 0.001

PlantHydraulics.critical_pressure — Method

critical_pressure(vc::LogisticVC{FT}, kr::FT = FT(0.001)) where {FT<:AbstractFloat}

Return the critical xylem water pressure at 25 °C that triggers a given amount of loss of conductance, given

vc ClimaCache.LogisticVC type struct

kr Reference conductance, default is 0.001

PlantHydraulics.critical_pressure — Method

critical_pressure(vc::PowerVC{FT}, kr::FT = FT(0.001)) where {FT<:AbstractFloat}

Return the critical xylem water pressure at 25 °C that triggers a given amount of loss of conductance, given

vc ClimaCache.PowerVC type struct

kr Reference conductance, default is 0.001

PlantHydraulics.critical_pressure — Method

critical_pressure(vc::WeibullVC{FT}, kr::FT = FT(0.001)) where {FT<:AbstractFloat}

Return the critical xylem water pressure at 25 °C that triggers a given amount of loss of conductance, given

vc ClimaCache.WeibullVC type struct

kr Reference conductance, default is 0.001

Pressure volume curve

PlantHydraulics.xylem_pressure — Function

This function returns the xylem water pressure from pressure volume curve. The supported methods are

xylem_pressure(pv, rvol, T)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_volume.jl:34.

xylem_pressure(pv, rvol, T)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_volume.jl:53.

PlantHydraulics.xylem_pressure — Method

xylem_pressure(pv::LinearPVCurve{FT}, rvol::FT, T::FT) where {FT<:AbstractFloat}

Return the xylem water pressure in MPa, given

pv LinearPVCurve type pressure volume curve

rvol Relative volume (relative to maximum)

T Temperature

PlantHydraulics.xylem_pressure — Method

xylem_pressure(pv::SegmentedPVCurve{FT}, rvol::FT, T::FT) where {FT<:AbstractFloat}

Return the xylem water pressure in MPa, given

pv SegmentedPVCurve type pressure volume curve

rvol Relative volume (relative to maximum)

T Temperature

PlantHydraulics.capacitance_buffer — Function

This function returns the xylem water buffer rate from pressure volume curve. The supported methods are

capacitance_buffer(pvc)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_volume.jl:99.

capacitance_buffer(pvc)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_volume.jl:118.

PlantHydraulics.capacitance_buffer — Method

capacitance_buffer(pvc::LinearPVCurve{FT}) where {FT<:AbstractFloat}

Return the relative capacictance buffer rate, given

pv LinearPVCurve type pressure volume curve

PlantHydraulics.capacitance_buffer — Method

capacitance_buffer(pvc::SegmentedPVCurve{FT}) where {FT<:AbstractFloat}

Return the relative capacictance buffer rate, given

pv SegmentedPVCurve type pressure volume curve

Cavitation legacy

PlantHydraulics.clear_legacy! — Function

This function clear the legacy for the hydraulic system. Supported methods are

clear_legacy!(hs)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/legacy.jl:31.

clear_legacy!(organ)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/legacy.jl:54.

clear_legacy!(spac)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/legacy.jl:71.

clear_legacy!(spac)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/legacy.jl:94.

clear_legacy!(spac)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/legacy.jl:116.

clear_legacy!(spac)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/legacy.jl:139.

PlantHydraulics.clear_legacy! — Method

clear_legacy!(hs::Union{LeafHydraulics{FT}, RootHydraulics{FT}, StemHydraulics{FT}}) where {FT<:AbstractFloat}

Clear the legacy for hydraulic system, given

hs LeafHydraulics, RootHydraulics, or StemHydraulics type structure

PlantHydraulics.clear_legacy! — Method

clear_legacy!(organ::Union{Leaf{FT}, Root{FT}, Stem{FT}}) where {FT<:AbstractFloat}

Clear the legacy for hydraulic system, given

organ Leaf, Root, or Stem type structure

PlantHydraulics.clear_legacy! — Method

clear_legacy!(spac::MonoElementSPAC{FT}) where {FT<:AbstractFloat}

Clear the legacy for SPAC system, given

spac MonoElementSPAC type structure

PlantHydraulics.clear_legacy! — Method

clear_legacy!(spac::MonoGrassSPAC{FT}) where {FT<:AbstractFloat}

Clear the legacy for SPAC system, given

spac MonoGrassSPAC type structure

PlantHydraulics.clear_legacy! — Method

clear_legacy!(spac::MonoPalmSPAC{FT}) where {FT<:AbstractFloat}

Clear the legacy for SPAC system, given

spac MonoPalmSPAC type structure

PlantHydraulics.clear_legacy! — Method

clear_legacy!(spac::MonoTreeSPAC{FT}) where {FT<:AbstractFloat}

Clear the legacy for SPAC system, given

spac MonoTreeSPAC type structure

FLow profile

PlantHydraulics.xylem_flow — Function

This function returns the sum flow rate of the downstream organs. The supported methods are

xylem_flow(mode)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:31.

xylem_flow(mode)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:48.

xylem_flow(hs)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:65.

xylem_flow(organ)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:82.

xylem_flow(organs)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:99.

xylem_flow(organs)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:123.

PlantHydraulics.xylem_flow — Method

xylem_flow(mode::SteadyStateFlow{FT}) where {FT<:AbstractFloat}

Return the flow rate, given

mode SteadyStateFlow type flow profile

PlantHydraulics.xylem_flow — Method

xylem_flow(mode::NonSteadyStateFlow{FT}) where {FT<:AbstractFloat}

Return the flow rate, given

mode NonSteadyStateFlow type flow profile

PlantHydraulics.xylem_flow — Method

xylem_flow(hs::Union{LeafHydraulics{FT}, RootHydraulics{FT}, StemHydraulics{FT}}) where {FT<:AbstractFloat}

Return the flow rate, given

hs LeafHydraulics, RootHydraulics, or StemHydraulics type struct

PlantHydraulics.xylem_flow — Method

xylem_flow(organ::Union{Leaf{FT}, Root{FT}, Stem{FT}}) where {FT<:AbstractFloat}

Return the flow rate, given

organ Leaf, Root, or Stem type struct

PlantHydraulics.xylem_flow — Method

xylem_flow(organs::Vector{Leaf{FT}}) where {FT<:AbstractFloat}

Return the sum flow rate, given

organs Vector of Leaf type struct

PlantHydraulics.xylem_flow — Method

xylem_flow(organs::Vector{Stem{FT}}) where {FT<:AbstractFloat}

Return the sum flow rate, given

organs Vector of Stem type struct

PlantHydraulics.root_pk — Function

This function returns the root end pressure and total hydraulic conductance to find flow rates in all roots. The supported methods are

root_pk(hs, mode, T)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:167.

root_pk(hs, mode, T)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:226.

root_pk(hs, T)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:284.

PlantHydraulics.root_pk — Method

root_pk(hs::RootHydraulics{FT}, mode::SteadyStateFlow{FT}, T::FT) where {FT<:AbstractFloat}

Return the root end pressre and total hydraulic conductance at steady state mode, given

hs RootHydraulics type struct

mode SteadyStateFlow type steady state flow

T Liquid temperature

PlantHydraulics.root_pk — Method

root_pk(hs::RootHydraulics{FT}, mode::NonSteadyStateFlow{FT}, T::FT) where {FT<:AbstractFloat}

Return the root end pressre and total hydraulic conductance at steady state mode, given

hs RootHydraulics type struct

mode NonSteadyStateFlow type non-steady state flow

T Liquid temperature

PlantHydraulics.root_pk — Method

root_pk(hs::RootHydraulics{FT}, T::FT) where {FT<:AbstractFloat}

Return the root end pressre and total hydraulic conductance at steady state mode, given

hs RootHydraulics type struct

T Liquid temperature

PlantHydraulics.xylem_flow_profile! — Function

This function update the flow rate profiles within the plant hydraulic system. The supported methods are

xylem_flow_profile!(mode, flow)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:323.

xylem_flow_profile!(mode, f_out)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:348.

xylem_flow_profile!(hs, mode, T, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:374.

xylem_flow_profile!(hs, mode, T, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:395.

xylem_flow_profile!(hs, mode, T, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:437.

xylem_flow_profile!(hs, T, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:483.

xylem_flow_profile!(organ, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:502.

xylem_flow_profile!(roots, cache_f, cache_k, cache_p, f_sum, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:526.

xylem_flow_profile!(spac, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:584.

xylem_flow_profile!(spac, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:617.

xylem_flow_profile!(spac, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:646.

xylem_flow_profile!(spac, Δt)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/flow_profile.jl:679.

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(mode::SteadyStateFlow, flow::FT) where {FT<:AbstractFloat}

Set up flow rate out for root at steady state mode, given

hs LeafHydraulics, RootHydraulics, or StemHydraulics type struct

mode SteadyStateFlow type steady state flow

flow Target flow rate (guess or solution)

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(mode::NonSteadyStateFlow, f_out::FT) where {FT<:AbstractFloat}

Set up flow rate out for root at non-steady state mode, given

hs LeafHydraulics, RootHydraulics, or StemHydraulics type struct

mode SteadyStateFlow type steady state flow

f_out Target flow rate out of the root (guess or solution)

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(hs::Union{LeafHydraulics{FT}, RootHydraulics{FT}, StemHydraulics{FT}}, mode::SteadyStateFlow{FT}, T::FT, Δt::FT) where {FT<:AbstractFloat}

Update organ flow rate profile at steady state mode after setting up the flow rate out (do nothing here), given

hs LeafHydraulics, RootHydraulics, or StemHydraulics type struct

mode SteadyStateFlow type steady state flow

T Liquid temperature

Δt Time step length

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(hs::LeafHydraulics{FT}, mode::NonSteadyStateFlow{FT}, T::FT, Δt::FT) where {FT<:AbstractFloat}

Update leaf flow rate at non-steady state mode after setting up the flow rate out, given

hs LeafHydraulics type struct

mode NonSteadyStateFlow type non-steady state flow

T Liquid temperature

Δt Time step length

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(hs::Union{RootHydraulics{FT}, StemHydraulics{FT}}, mode::NonSteadyStateFlow{FT}, T::FT, Δt::FT) where {FT<:AbstractFloat}

Update root or stem flow rate at non-steady state mode after setting up the flow rate out, given

hs RootHydraulics or StemHydraulics type struct

mode NonSteadyStateFlow type non-steady state flow

T Liquid temperature

Δt Time step length

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(hs::Union{LeafHydraulics{FT}, RootHydraulics{FT}, StemHydraulics{FT}}, T::FT, Δt::FT) where {FT<:AbstractFloat}

Update root or stem flow rate at non-steady state mode after setting up the flow rate out, given

hs LeafHydraulics, RootHydraulics, or StemHydraulics type struct

T Liquid temperature

Δt Time step length

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(organ::Union{Leaf{FT}, Root{FT}, Stem{FT}}, Δt::FT) where {FT<:AbstractFloat}

Update root or stem flow rate at non-steady state mode after setting up the flow rate out, given

hs Leaf, Root, or Stem type struct

Δt Time step length

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(roots::Vector{Root{FT}}, cache_f::Vector{FT}, cache_k::Vector{FT}, cache_p::Vector{FT}, f_sum::FT, Δt::FT) where {FT<:AbstractFloat}

Update root or stem flow rate at non-steady state mode after setting up the flow rate out, given

roots Vector of Root in a multiple roots system

cache_f Flow rate cache into each root

cache_k Total conductance cache of each root

cache_p Root xylem end pressure cache of each root

f_sum Total

Δt Time step length

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(spac::MonoElementSPAC{FT}, Δt::FT) where {FT<:AbstractFloat}

Update flow profiles for the soil-plant-air continuum (after setting up leaf flow rate), given

spac MonoElementSPAC type SPAC system

Δt Time step length

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(spac::MonoGrassSPAC{FT}, Δt::FT) where {FT<:AbstractFloat}

Update flow profiles for the soil-plant-air continuum (after setting up leaf flow rate), given

spac MonoGrassSPAC type SPAC system

Δt Time step length

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(spac::MonoPalmSPAC{FT}, Δt::FT) where {FT<:AbstractFloat}

Update flow profiles for the soil-plant-air continuum (after setting up leaf flow rate), given

spac MonoPalmSPAC type SPAC system

Δt Time step length

PlantHydraulics.xylem_flow_profile! — Method

xylem_flow_profile!(spac::MonoTreeSPAC{FT}, Δt::FT) where {FT<:AbstractFloat}

Update flow profiles for the soil-plant-air continuum (after setting up leaf flow rate), given

spac MonoTreeSPAC type SPAC system

Δt Time step length

Pressure profile

PlantHydraulics.xylem_end_pressure — Function

This function returns the xylem end water pressure. The supported methods are

xylem_end_pressure(hs, flow, T)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:34.

xylem_end_pressure(hs, flow, T)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:76.

xylem_end_pressure(hs, flow, T)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:130.

xylem_end_pressure(organ, flow)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:171.

xylem_end_pressure(spac, flow)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:189.

xylem_end_pressure(spac, f_sl, f_sh, r_sl)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:219.

PlantHydraulics.xylem_end_pressure — Method

xylem_end_pressure(hs::LeafHydraulics{FT}, flow::FT, T::FT) where {FT<:AbstractFloat}

Return the xylem end water pressure in MPa, given

hs LeafHydraulics type struct

flow Flow rate per leaf area [mol m⁻² s⁻¹]

T Temperature

PlantHydraulics.xylem_end_pressure — Method

xylem_end_pressure(hs::RootHydraulics{FT}, flow::FT, T::FT) where {FT<:AbstractFloat}

Return the xylem end water pressure in MPa, given

hs RootHydraulics type struct

flow Flow rate [mol s⁻¹]

T Temperature

PlantHydraulics.xylem_end_pressure — Method

xylem_end_pressure(hs::StemHydraulics{FT}, flow::FT, T::FT) where {FT<:AbstractFloat}

Return the xylem end water pressure in MPa, given

hs StemHydraulics type struct

flow Flow rate [mol s⁻¹]

T Temperature

PlantHydraulics.xylem_end_pressure — Method

xylem_end_pressure(organ::Union{Leaf{FT}, Root{FT}, Stem{FT}}, flow::FT) where {FT<:AbstractFloat}

Return the xylem end water pressure in MPa, given

organ Leaf, Root, or Stem type struct

flow Flow rate (per leaf area for Leaf) [mol (m⁻²) s⁻¹]

PlantHydraulics.xylem_end_pressure — Method

xylem_end_pressure(spac::MonoElementSPAC{FT}, flow::FT) where {FT<:AbstractFloat}

Return the xylem end water pressure in MPa, given

spac MonoElementSPAC type struct

flow Flow rate [mol s⁻¹]

PlantHydraulics.xylem_end_pressure — Method

xylem_end_pressure(spac::MonoElementSPAC{FT}, f_sl::FT, f_sh::FT, r_sl::FT) where {FT<:AbstractFloat}

Return the xylem end water pressure in MPa for sunlit and shaded leaves, given

spac MonoElementSPAC type struct

f_sl Sunlit leaves flow rate [mol s⁻¹]

f_sh Shaded leaves flow rate [mol s⁻¹]

r_sl Sunlit leaves fraction

PlantHydraulics.xylem_pressure_profile! — Function

This function updates the xylem water pressure profile. The supported methods are

xylem_pressure_profile!(hs, mode, T; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:270.

xylem_pressure_profile!(hs, mode, T; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:328.

xylem_pressure_profile!(hs, mode, T; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:384.

xylem_pressure_profile!(hs, mode, T; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:449.

xylem_pressure_profile!(hs, mode, T; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:516.

xylem_pressure_profile!(hs, mode, T; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:569.

xylem_pressure_profile!(organ; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:621.

xylem_pressure_profile!(spac; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:639.

xylem_pressure_profile!(spac; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:668.

xylem_pressure_profile!(spac; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:705.

xylem_pressure_profile!(spac; update)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/pressure_profile.jl:746.

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(hs::LeafHydraulics{FT}, mode::SteadyStateFlow{FT}, T::FT; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

hs LeafHydraulics type hydraulic system

mode SteadyStateFlow struct with steady state flow

T Water temperature

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(hs::LeafHydraulics{FT}, mode::NonSteadyStateFlow{FT}, T::FT; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

hs LeafHydraulics type hydraulic system

mode NonSteadyStateFlow struct with steady state flow

T Water temperature

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(hs::RootHydraulics{FT}, mode::SteadyStateFlow{FT}, T::FT; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

hs RootHydraulics type hydraulic system

mode SteadyStateFlow struct with steady state flow

T Water temperature

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(hs::RootHydraulics{FT}, mode::NonSteadyStateFlow{FT}, T::FT; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

hs RootHydraulics type hydraulic system

mode NonSteadyStateFlow struct with steady state flow

T Water temperature

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(hs::StemHydraulics{FT}, mode::SteadyStateFlow{FT}, T::FT; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

hs StemHydraulics type hydraulic system

mode SteadyStateFlow struct with steady state flow

T Water temperature

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(hs::StemHydraulics{FT}, mode::NonSteadyStateFlow{FT}, T::FT; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

hs StemHydraulics type hydraulic system

mode NonSteadyStateFlow struct with steady state flow

T Water temperature

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(organ::Union{Leaf{FT}, Root{FT}, Stem{FT}}; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

organ Leaf, Root, or Stem type organ

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(spac::MonoElementSPAC{FT}; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

spac MonoElementSPAC type organ

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(spac::MonoGrassSPAC{FT}; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

spac MonoGrassSPAC type organ

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(spac::MonoPalmSPAC{FT}; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

spac MonoPalmSPAC type organ

update If true, update xylem cavitation legacy

PlantHydraulics.xylem_pressure_profile! — Method

xylem_pressure_profile!(spac::MonoTreeSPAC{FT}; update::Bool = true) where {FT<:AbstractFloat}

Update xylem pressure profile (flow profile needs to be updated a priori), given

spac MonoTreeSPAC type organ

update If true, update xylem cavitation legacy

Critical flow

PlantHydraulics.critical_flow — Function

This function returns the critical flow rate that triggers 99.triggers a given amount of loss of hydraulic conductance. The supported methods are

critical_flow(hs, T)
critical_flow(hs, T, ini; kr)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/target_flow.jl:34.

critical_flow(spac)
critical_flow(spac, ini; kr)

defined at /home/runner/work/PlantHydraulics.jl/PlantHydraulics.jl/src/target_flow.jl:80.

PlantHydraulics.critical_flow — Method

critical_flow(hs::LeafHydraulics{FT}, T::FT, ini::FT = FT(0.5); kr::FT = FT(0.001)) where {FT<:AbstractFloat}

Return the critical flow rate that triggers a given amount of loss of conductance, given

hs LeafHydraulics type struct

T Liquid temperature

ini Initial guess

kr Reference conductance, default is 0.001

PlantHydraulics.critical_flow — Method

critical_flow(spac::MonoElementSPAC{FT}, ini::FT = FT(0.5); kr::FT = FT(0.001)) where {FT<:AbstractFloat}

Return the critical flow rate that triggers a given amount of loss of conductance, given

spac MonoElementSPAC type struct

ini Initial guess

kr Reference conductance, default is 0.001