API

Structs

PlantHydraulics.AbstractHydraulicSystem — Type

type AbstractHydraulicSystem

Hierarchy of AbstractHydraulicSystem

LeafHydraulics

RootHydraulics

StemHydraulics

PlantHydraulics.LeafHydraulics — Type

struct LeafHydraulics{FT<:AbstractFloat}

A struct that contains leaf hydraulics information.

Fields

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_ox

Maximal extra-xylary hydraulic conductance [mol s⁻¹ MPa⁻¹ m⁻²]

k_sla

Maximal leaf hydraulic conductance per leaf area [mol s⁻¹ MPa⁻¹ m⁻²]

p_crt

Critical xylem pressure [MPa]

flow

Flow rate in the xylem [mol s⁻¹]

p_dos

Leaf xylem water pressure at the downstream end of leaf xylem [MPa]

p_leaf

Leaf end water pressure at air-water interface [MPa]

p_ups

Leaf xylem water pressure at the leaf base (upstream) [MPa]

k_element

List of leaf k_max per element [mol s⁻¹ MPa⁻¹ m⁻²]

k_history

List of leaf kr history per element

p_element

List of xylem water pressure [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) [MPa]

f_st

Relative surface tension

f_vis

Relative viscosity

T_sap

Upstream sap temperature [K]

PlantHydraulics.RootHydraulics — Type

struct RootHydraulics{FT<:AbstractFloat}

A struct that contains root hydraulics information.

Fields

area

Root cross-section area [m²]

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_max

Maximal hydraulic conductance [mol s⁻¹ MPa⁻¹]

k_s

Maximal xylem hydraulic conductivity [mol s⁻¹ MPa⁻¹ m⁻²]

Δh

Root z difference [m]

k_rhiz

Rhizosphere conductance [mol s⁻¹ MPa⁻¹]

soil_type

Soil type

soil_α

α is related to the inverse of the air entry suction, α > 0

soil_n

Measure of the pore-size distribution

soil_m

1 - 1/n

flow

Flow rate in the xylem [mol s⁻¹]

p_dos

Xylem water pressure at the downstream end of xylem [MPa]

p_rhiz

Xylem-rhizosphere interface water pressure [MPa]

p_ups

Soil matrix potential [MPa]

k_element

List of k_max per element [mol s⁻¹ MPa⁻¹ m⁻²]

k_history

List of kr history per element

p_element

List of xylem water pressure [MPa]

p_gravity

List of pressure drop caused by gravity [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) [MPa]

f_st

Relative surface tension

f_vis

Relative viscosity

T_sap

Upstream sap temperature [K]

PlantHydraulics.StemHydraulics — Type

struct StemHydraulics{FT<:AbstractFloat}

A struct that contains stem hydraulics information.

Fields

area

Stem cross-section area [m²]

b

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter B [MPa]

c

Weibull function (k = k_max * exp( -(-p/B)^C )) parameter C

k_max

Maximal hydraulic conductance [mol s⁻¹ MPa⁻¹]

k_s

Maximal xylem hydraulic conductivity [mol s⁻¹ MPa⁻¹ m⁻²]

Δh

Stem height difference [m]

p_dos

Xylem water pressure at the downstream end of xylem [MPa]

p_ups

Xylem water pressure at the base (upstream) [MPa]

k_element

List of k_max per element [mol s⁻¹ MPa⁻¹ m⁻²]

k_history

List of kr history per element

p_element

List of xylem water pressure [MPa]

p_gravity

List of pressure drop caused by gravity [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) [MPa]

f_st

Relative surface tension

f_vis

Relative viscosity

T_sap

Upstream sap temperature [K]

Xylem hydraulics

PlantHydraulics.weibull_k_ratio — Function

weibull_k_ratio(b::FT, c::FT, p_25::FT, vis::FT)
weibull_k_ratio(b::FT, c::FT, p::FT)

Returns the relative hydraulic conductance, given

b Weibull B in [MPa]

c Weibull C

p_25 Equivalent xylem pressure at 298.15 K in [MPa]

vis Relative viscosity. If missing, vis = 1.

PlantHydraulics.xylem_p_from_flow — Function

xylem_p_from_flow(hs::LeafHydraulics{FT}, flow::FT)

Return the xylen end pressure from flow rate, given

hs AbstractHydraulicSystem type struct

flow Flow rate (per leaf area)

Note, gravity is accounted for in root and stem; rhizosphere conductance is accounted for in root; extra-xylary conductance is not accounted for in leaf here because it calculates xylem end pressure.

PlantHydraulics.root_q_from_pressure — Function

root_q_from_pressure(root::RootHydraulics, pressure::FT, ini::FT)

Calculate the flow rate from a given tree base pressure, given

root RootHydraulics type struct

pressure Given tree base pressure in [MPa]

ini Start point for RootSolvers

PlantHydraulics.q_diff — Function

q_diff(roots::Array{RootHydraulics{FT},1}, pressure::FT, target::FT)

Function to calculate the flow rate difference to find solution pressure, given

root RootHydraulics type struct

pressure Given tree base pressure in [MPa]

target Target flow rate in [mol s⁻¹]

PlantHydraulics.root_qs_p_from_q — Function

root_qs_p_from_q(roots::Array{RootHydraulics{FT},1}, flow::FT, ini::FT)

Solve for the flow rates in each root layer and root pressure, given

roots Array of RootHydraulics struts

flow Total flow rates

ini Initial guess of plant base pressure in [MPa]

PlantHydraulics.hydraulic_p_profile! — Function

hydraulic_p_profile!(hs::LeafHydraulics{FT}, flow::FT)

Update the pressure profile, given

hs AbstractHydraulicSystem type struct

flow Flow rate (per leaf area)

PlantHydraulics.leaf_xylem_risk — Function

leaf_xylem_risk(hs::LeafHydraulics{FT}, flow::FT)

Evaluate the hydraulic risk at the end of leaf xylem, given

hs AbstractHydraulicSystem type struct

flow Flow rate (per leaf area)

PlantHydraulics.leaf_e_crit — Function

leaf_e_crit(hs::LeafHydraulics{FT})

Calculate the critical flow rate (K ≈ 0), given

hs AbstractHydraulicSystem type struct

Note, for the safety of no NaN, update e_crit when ΔP >= -0.01