API

Plant Hydraulic System

The PlantHydraulics module provides two levels of hydraulics system: organ-level and plant-level. The organ-level hydraulic systems include Leaf, Root, and Stem (trunk and branch). The plant-level hydraulic system is can be any combination of the three organs (custimized definition may apply).

Leaf, Root, and Stem HS

Plant hydraulics is segmented to three organ-level systems/structs (LeafHydraulics, RootHydraulics, and StemHydraulics) subject to an Abstract type (AbstractHydraulicSystem). The major differences among the three structs are

LeafHydraulics has an extra-xylary component

RootHydraulics has a rhizosphere component

RootHydraulics and StemHydraulics have a gravity component

See the documentation for each struct for more details:

PlantHydraulics.AbstractHydraulicSystem — Type

abstract type AbstractHydraulicSystem{FT}

Hierarchy of AbstractHydraulicSystem

LeafHydraulics

RootHydraulics

StemHydraulics

PlantHydraulics.LeafHydraulics — Type

mutable struct LeafHydraulics{FT<:AbstractFloat}

A struct that contains leaf hydraulics information.

Fields

area

Leaf area [m²]

k_ox

Maximal extra-xylary hydraulic conductance [mol s⁻¹ MPa⁻¹ m⁻²]

k_sla

Maximal leaf hydraulic conductance per leaf area [mol s⁻¹ MPa⁻¹ m⁻²]

vc

Vulnerability curve

p_crt

Critical xylem pressure [MPa]

flow

Flow rate in the xylem [mol s⁻¹]

p_dos

Leaf xylem water pressure at the downstream end of leaf xylem [MPa]

p_leaf

Leaf end water pressure at air-water interface [MPa]

p_ups

Leaf xylem water pressure at the leaf base (upstream) [MPa]

k_element

List of leaf k_max per element [mol s⁻¹ MPa⁻¹ m⁻²]

k_history

List of leaf kr history per element

p_element

List of xylem water pressure [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) [MPa]

f_st

Relative surface tension

f_vis

Relative viscosity

T_sap

Upstream sap temperature [K]

PlantHydraulics.RootHydraulics — Type

mutable struct RootHydraulics{FT<:AbstractFloat}

A struct that contains root hydraulics information.

Fields

area

Root cross-section area [m²]

k_max

Maximal hydraulic conductance [mol s⁻¹ MPa⁻¹]

k_s

Maximal xylem hydraulic conductivity [mol s⁻¹ MPa⁻¹ m⁻²]

vc

Vulnerability curve

Δh

Root z difference [m]

k_rhiz

Rhizosphere conductance [mol s⁻¹ MPa⁻¹]

soil_type

Soil type

soil_α

α is related to the inverse of the air entry suction, α > 0

soil_n

Measure of the pore-size distribution

soil_m

1 - 1/n

flow

Flow rate in the xylem [mol s⁻¹]

p_dos

Xylem water pressure at the downstream end of xylem [MPa]

p_rhiz

Xylem-rhizosphere interface water pressure [MPa]

p_ups

Soil matrix potential [MPa]

k_element

List of k_max per element [mol s⁻¹ MPa⁻¹ m⁻²]

k_history

List of kr history per element

p_element

List of xylem water pressure [MPa]

p_gravity

List of pressure drop caused by gravity [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) [MPa]

f_st

Relative surface tension

f_vis

Relative viscosity

T_sap

Upstream sap temperature [K]

PlantHydraulics.StemHydraulics — Type

mutable struct StemHydraulics{FT<:AbstractFloat}

A struct that contains stem hydraulics information.

Fields

area

Stem cross-section area [m²]

k_max

Maximal hydraulic conductance [mol s⁻¹ MPa⁻¹]

k_s

Maximal xylem hydraulic conductivity [mol s⁻¹ MPa⁻¹ m⁻²]

vc

Vulnerability curve

Δh

Stem height difference [m]

p_dos

Xylem water pressure at the downstream end of xylem [MPa]

p_ups

Xylem water pressure at the base (upstream) [MPa]

k_element

List of k_max per element [mol s⁻¹ MPa⁻¹ m⁻²]

k_history

List of kr history per element

p_element

List of xylem water pressure [MPa]

p_gravity

List of pressure drop caused by gravity [MPa]

p_history

List of xylem water pressure history (normalized to 298.15 K) [MPa]

f_st

Relative surface tension

f_vis

Relative viscosity

T_sap

Upstream sap temperature [K]

To initialize a hydraulics system, one needs to provide the floating type, for example:

using PlantHydraulics

FT = Float32;
hs_leaf = LeafHydraulics{FT}();
hs_root = RootHydraulics{FT}();
hs_stem = StemHydraulics{FT}();

Whole-plant HS

Plants differ in their structures, for example, some plants have a canopy far above the ground elevated by a trunk, some plants have a structured canopy supported by branch systems, and some plant has no trunk at all. To represent the structural differences, several types of plant hydraulics systems are pre-defined, and they are GrassLikeHS, PalmLikeHS, and TreeLikeHS structs subject to a AbstractPlantHS type, where HS stands for hydraulic system. The major difference between the HSs are

GrassLikeHS has only mutiple root and canopy layers, no trunk or branch

PalmLikeHS has multiple root layers, a trunk, and multiple canopy layers, no branch system

TreeLikeHS has multiple root layers, a trunk, and multiple branch+canopy layers, and each branch corresponds to a canopy layer

See the documentation for each struct for more details:

PlantHydraulics.AbstractPlantHS — Type

abstract type AbstractPlantHS{FT}

Hierachy of AbstractPlantHS

GrassLikeHS

PalmLikeHS

TreeLikeHS

PlantHydraulics.GrassLikeHS — Type

mutable struct GrassLikeHS{FT<:AbstractFloat}

A plant hydraulic system like a grass, which contains multiple root layers, and multiple canopy layers. No trunk or branch system applies.

Fields

n_root

Root Layers

n_canopy

Canopy Layers

roots

Roots system

leaves

Leaves

root_index_in_soil

Corresponding soil layer per root layer

canopy_index_in_air

Corresponding air layer per canopy layer

PlantHydraulics.PalmLikeHS — Type

mutable struct PalmLikeHS{FT<:AbstractFloat}

A plant hydraulic system like a palm, which contains multiple root layers, one trunk, and multiple canopy layers. No branch system applies.

Fields

n_root

Root Layers

n_canopy

Canopy Layers

roots

Roots system

trunk

Trunk

leaves

Leaves

root_index_in_soil

Corresponding soil layer per root layer

canopy_index_in_air

Corresponding air layer per canopy layer

PlantHydraulics.TreeLikeHS — Type

mutable struct TreeLikeHS{FT<:AbstractFloat}

A plant hydraulic system like a tree, which contains multiple root layers, one trunk, and multiple branch and canopy layers.

Fields

n_root

Root Layers

n_canopy

Canopy Layers

roots

Roots system

trunk

Trunk

branch

Branch system

leaves

Leaves

root_index_in_soil

Corresponding soil layer per root layer

canopy_index_in_air

Corresponding air layer per canopy layer

To ease the initialization of a plant hydraulics system, a few customized functions are provided for quick initialization. More importantly, modifications to each field in the struct are always allowed. The quick functions are create_grass_like_hs, create_palm_like_hs, and create_tree_like_hs:

PlantHydraulics.create_grass_like_hs — Function

create_grass_like_hs(z_root::FT, z_trunk::FT, z_canopy::FT, soil_bounds::Array{FT,1}, air_bounds::Array{FT,1}) where {FT<:AbstractFloat}

Create a GrassLikeHS, given

z_root Maximal root depth (negative value)

z_canopy Maximal canopy height (positive value)

soil_bounds Array of soil layer boundaries starting from 0

air_bounds Array of air layer boundaries starting from 0

PlantHydraulics.create_palm_like_hs — Function

create_palm_like_hs(z_root::FT, z_trunk::FT, z_canopy::FT, soil_bounds::Array{FT,1}, air_bounds::Array{FT,1}) where {FT<:AbstractFloat}

Create a PalmLikeHS, given

z_root Maximal root depth (negative value)

z_trunk Maximal trunk height (positive value)

z_canopy Maximal canopy height (positive value)

soil_bounds Array of soil layer boundaries starting from 0

air_bounds Array of air layer boundaries starting from 0

PlantHydraulics.create_tree_like_hs — Function

create_tree_like_hs(z_root::FT, z_trunk::FT, z_canopy::FT, soil_bounds::Array{FT,1}, air_bounds::Array{FT,1}) where {FT<:AbstractFloat}

Create a TreeLikeHS, given

z_root Maximal root depth (negative value)

z_trunk Maximal trunk height (positive value)

z_canopy Maximal canopy height (positive value)

soil_bounds Array of soil layer boundaries starting from 0

air_bounds Array of air layer boundaries starting from 0

What these functions do are to determine how many root layers and branch/canopy layers to add based on the tree information and environmental settings. To determine number of root layers, rooting depth and the soil layer information are required. The z_root is the maximal root depth in negative number, and soil_bounds is the boundaries of soil layers staring from 0. For example, for a soil_bounds of [0.0, -1.0, -2.0, -3.0, -4.0], a z_root of -1 gives 1 root layer, and a z_root of -1.5 or -2.0 gives 2 root layers. The z_trunk, z_canopy, and air_bounds determine how many canopy layers to add. For example, for a air_bounds of [0.0, 1.0, 2.0, 3.0, 4.0, 5.0 ... 20.0, 21.0, 22.0], a z_trunk of 5.0 z_canopy of 7.0 give 2 canopy layers, and a z_trunk of 5.5 z_canopy of 7.0 give 2 canopy layers. Also, the root_index_in_soil and canopy_index_in_air indicate which soil or air layer the root or canopy layer corresponds with, respectively. For instance, a index of 7 means that the canopy layer should use the 7th air layer.

To initialize a whole-plant hydraulic system, checkout the example below:

using PlantHydraulics

FT = Float32;
grass = create_grass_like_hs(FT(-2.1), FT(0.5), FT(8), FT[0,-1,-2,-3], collect(FT,0:1:20));
palm  =  create_palm_like_hs(FT(-2.1), FT(0.5), FT(8), FT[0,-1,-2,-3], collect(FT,0:1:20));
tree  =  create_tree_like_hs(FT(-2.1), FT(0.5), FT(8), FT[0,-1,-2,-3], collect(FT,0:1:20));

Hydraulic conductance

Plants transport water through xylem conduits (vessels in most angiosperms, trachieds in most gymnosperms). With the ascent of sap along the hydraulic system, water pressure in the conduits is typically negative. The negative xylem water pressure tends to pull air from surrounding tisses or the atmosphere into the xylem conduits, resulting in xylem cavitation. The air bubbles in cavitated conduits block water flow, and thus results in decline of water transport capability (measured by xylem hydraulic conductance).

Typically, the correlation between xylem water pressure ($P \leq 0$) and hydraulic conductance ($k$) is expressed by a Weibull function for WeibullSingle type correlation:

\[k = k_\text{max} \cdot \exp \left( -\left( \dfrac{-P}{B} \right)^C \right)\]

where $k_\text{max}$ is the maximal hydraulic conductance, and $B$ and $C$ are the Weibull parameters. This correlation is also known as vulnerability curve (VC) to drought stress. Sometimes, plants exhibit a segmented VC, for example, the fibers may transport water as well and are much more resistant to drought than vessels. Thus, a dual Weibull function is presented for WeibullDual type correlation ($P \leq 0$):

\[k = k_\text{max} \cdot \left\{ f_1 \cdot \exp \left[ -\left( \dfrac{-P}{B_1} \right)^{C_1} \right] + (1 - f_1) \cdot \exp \left[ -\left( \dfrac{-P}{B_2} \right)^{C_2} \right] \right\}\]

The VC formulations are abstractized as

PlantHydraulics.AbstractVulnerability — Type

abstract type AbstractVulnerability{FT}

Hierachy of AbstractVulnerability

WeibullSingle

WeibullDual

PlantHydraulics.WeibullDual — Type

mutable struct WeibullDual{FT}

Struct that contains dual Weibull function parameters.

Fields

b1

B of first part [MPa]

c1

C of first part

f1

F of first part

b2

B of second part [MPa]

c2

C of second part

f2

F of second part

PlantHydraulics.WeibullSingle — Type

mutable struct WeibullSingle{FT}

Struct that contains single Weibull function parameters.

Fields

b

B of first part [MPa]

c

C of first part

The function to call is

PlantHydraulics.xylem_k_ratio — Function

xylem_k_ratio(vc::WeibullSingle{FT}, p_25::FT, vis::FT) where {FT<:AbstractFloat}
xylem_k_ratio(vc::  WeibullDual{FT}, p_25::FT, vis::FT) where {FT<:AbstractFloat}

Returns the relative hydraulic conductance, given

vc Xylem vulnerability curve

p Xylem pressure at 298.15 K in [MPa]

p_25 Equivalent xylem pressure at 298.15 K in [MPa]

vis Relative viscosity. If missing, vis = 1.

Note it here that xylem_k_ratio(vc, p) calculate the k without making temperature corrections, but the xylem_k_ratio(vc, p_25, vis) makes correction over the viscosity (the higher the viscosity, the lower the k). Also, p_25 means that the pressure has been corrected to 298.15 K for surface tension (the higher the surface tension, the more resistant the xylem).

Meanwhile, there is a function to call to calculate the critical pressure, beyond which leaf will decicate. The critical pressure is calculated as the pressure at which $k$ is 0.001 of $k_\text{max}$ for WeibullSingle (for WeibullDual, each segment need to reach 0.001). The functions is

PlantHydraulics.xylem_p_crit — Function

xylem_p_crit(vc::WeibullSingle{FT}, st::FT) where {FT<:AbstractFloat}
xylem_p_crit(vc::WeibullDual{FT}  , st::FT) where {FT<:AbstractFloat}

Returns the relative hydraulic conductance, given

vc Xylem vulnerability curve

st Relative surface tension. If missing, vis = 1.

Examples:

using PlantHydraulics

FT = Float32;
vc_1 = WeibullSingle{FT}();
vc_2 = WeibullDual{FT}();

k_1 = xylem_k_ratio(vc_1, -1.0);
k_2 = xylem_k_ratio(vc_2, -1.0);
k_3 = xylem_k_ratio(vc_1, -1.0, 1.2);
k_4 = xylem_k_ratio(vc_2, -1.0, 1.2);

Pressure and Flow

The PlantHydraulics module is designed to run numerically for the following reasons:

Weibull function is cannot be integrated

The VC is segmented, i.e., if $P > 0$, $k = k_\text{max}$ (implemented in xylem_k_ratio)

Once xylem cavitation occurs, it cannot be easily recovered unless $P > 0$, and thus there is a drought legacy effect. This is why there are a few fields in the LeafHydraulics, RootHydraulics, and StemHydraulics structs to store the drought history information.

Temperature may change along the flow path. The f_st and f_vis in the structs help deal with these effects.

Function xylem_p_from_flow calculates the xylem end pressure for an organ-level hysraulic system. As mentioned above, the RootHydraulics and StemHydraulics has a gravity component, and the RootHydraulics has a rhizosphere component. Also be aware that xylem_p_from_flow accounts for temperature effects on surface tension and viscosity.

PlantHydraulics.xylem_p_from_flow — Function

xylem_p_from_flow(hs::LeafHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
xylem_p_from_flow(hs::RootHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
xylem_p_from_flow(hs::StemHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}

Return the xylen end pressure from flow rate, given

hs LeafHydraulics or RootHydraulics or StemHydraulics type struct

flow Flow rate (per leaf area)

Note, gravity is accounted for in root and stem; rhizosphere conductance is accounted for in root; extra-xylary conductance is not accounted for in leaf here because it calculates xylem end pressure.

Noe that function xylem_p_from_flow does not update the pressure profiles or history in the xylem. To update these profiles, use hydraulic_p_profile!:

PlantHydraulics.hydraulic_p_profile! — Function

hydraulic_p_profile!(hs::LeafHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
hydraulic_p_profile!(hs::RootHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}
hydraulic_p_profile!(hs::StemHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}

Update the pressure profile, given

hs LeafHydraulics or RootHydraulics or StemHydraulics type struct

flow Flow rate (per leaf area)

Examples:

using PlantHydraulics

FT = Float32;
leaf = LeafHydraulics{FT}();
p = xylem_p_from_flow(leaf, FT(0.01));
@show leaf.p_element;
hydraulic_p_profile!(leaf, FT(0.01));
@show leaf.p_element;

Root Hydraulics

Function xylem_p_from_flow works for the case of only 1 root layer if one needs the plant base xylem water pressure. However, when there are multiple root layers, xylem_p_from_flow does not apply. In this case, iterations are required to calculate the xylem end pressure for each root layers, and then make sure all root layers have the same xylem end pressure. A few functions are provided to realize this.

Function root_q_from_pressure uses Root Solving method to calculate the flow rate through the RootHydraulics struct that yields the given xylem end pressure. The ini in the function is optional. However, using the flow rate from last instant when pressure does not differ much will speed up the calculation.

PlantHydraulics.root_q_from_pressure — Function

root_q_from_pressure(root::RootHydraulics{FT}, pressure::FT, ini::FT) where {FT<:AbstractFloat}

Calculate the flow rate from a given tree base pressure, given

root RootHydraulics type struct

pressure Given tree base pressure in [MPa]

ini Initial guess

Computing the flow rates in all the root layers and the plant base xylem pressure from a given flow needs more iterations based on root_q_from_pressure. The algorithm is to calculate the flow rates in all the root layers for a given plant base xylem pressure, and then use the Root Solving methods to calculate pressure that yields a total flow rate equals to the given value. The function q_diff calculates the difference of total flow rate for a given pressure and the target flow rate. The function root_qs_p_from_q calculates the flows rates in all the root layers and the plant base xylem pressure. Note that ini in the function is also optional, but a good guess will speed up the calculations. Be aware that function root_qs_p_from_q does not update the pressure profiles in the plant struct.

PlantHydraulics.q_diff — Function

q_diff(roots::Array{RootHydraulics{FT},1}, pressure::FT, target::FT) where {FT<:AbstractFloat}

Function to calculate the flow rate difference to find solution pressure, given

roots Array of RootHydraulics type struct

pressure Given tree base pressure in [MPa]

target Target flow rate in [mol s⁻¹]

PlantHydraulics.root_qs_p_from_q — Function

root_qs_p_from_q(roots::Array{RootHydraulics{FT},1}, flow::FT, ini::FT) where {FT<:AbstractFloat}

Solve for the flow rates in each root layer and root pressure, given

roots Array of RootHydraulics struts

flow Total flow rates

ini Initial guess of plant base pressure in [MPa]

Example

using PlantHydraulics

FT = Float32;
palm  =  create_palm_like_hs(FT(-2.1), FT(0.5), FT(8), FT[0,-1,-2,-3], collect(FT,0:1:20));
qs,p = root_qs_p_from_q(palm.roots, FT(1));

Leaf Hydraulics

The stomatal models often require plant hydraulics either as a correction factor (in empirical stomatal models) or as the risk term (in optimal stomatal models). To facilitate the calculations, a few specific functions are provided.

Function leaf_xylem_risk returns the risk in xylem hydraulic function based on the most downstream end of the xylem. The risk of plant hydraulic system is not only on current system, but also potential new growth (plants don't want to risk new growth either). Thus, function leaf_xylem_risk evaluates the risk from the xylem pressure calculated from current system (with drought history), and then compute the risk from the pressure (the severer the srought history, the higher the risk):

PlantHydraulics.leaf_xylem_risk — Function

leaf_xylem_risk(hs::LeafHydraulics{FT}, flow::FT) where {FT<:AbstractFloat}

Evaluate the hydraulic risk at the end of leaf xylem, given

hs LeafHydraulics type struct

flow Flow rate (per leaf area)

Note that function leaf_xylem_risk can work on its own without having other organ-level components. For example, by changing the p_ups of a LeafHydraulics, one can simulate the case of drought without caring about other hydraulic systems. Same for function leaf_e_crit below. However, these functions are only useful for sensitivity analysis or when p_ups in the LeafHydraulics is accurate.

Examples

using PlantHydraulics

FT = Float32;
leaf = LeafHydraulics{FT}();
risk = leaf_xylem_risk(leaf, FT(0.01));
@show risk;
leaf.p_ups = FT(-1.0);
risk = leaf_xylem_risk(leaf, FT(0.01));
@show risk;

Function leaf_e_crit calculates critical leaf transpiration rate, beyond which leaf will desicate. Function leaf_e_crit accounts for drought legacy effect by design, and the more severe the drought history, the lower the e_crit. Again, ini in the function is also optional, but a good guess will speed up the calculations.

PlantHydraulics.leaf_e_crit — Function

leaf_e_crit(hs::LeafHydraulics{FT}, ini::FT) where {FT<:AbstractFloat}

Calculate the critical flow rate (K ≈ 0), given

hs LeafHydraulics type struct

ini Initial guess

Note, for the safety of no NaN, update e_crit when ΔP >= -0.01

Examples

using PlantHydraulics

FT = Float32;
leaf = LeafHydraulics{FT}();
risk = leaf_e_crit(leaf);
@show risk;
leaf.p_ups = FT(-1.0);
risk = leaf_e_crit(leaf);
@show risk;

Whow-plant Hydraulics

Though leaf_xylem_risk and leaf_e_crit can work on their own, the functions only evaluate the risks on leaf level. The more realistic case is that when leaf transpiration rate increases, p_ups in the LeafHydraulics gets more negative. Thus, the leaf_xylem_risk and leaf_e_crit tends to underestimate the risk and overestimate the critical flow rate. To overcome this problem, whole-plant level plant hydraulics are provided.

Function tree_p_from_flow calculates the leaf xylem end pressure for a whole-plant struct using these steps:

calculate the plant base pressure from a given total flow rate

calculate the trunk end pressure (if present)

calculate the branch end pressure (if present)

calculate the leaf end pressure (if present)

PlantHydraulics.tree_p_from_flow — Function

tree_p_from_flow(tree::GrassLikeHS{FT}, flow::FT) where {FT<:AbstractFloat}
tree_p_from_flow(tree:: PalmLikeHS{FT}, flow::FT) where {FT<:AbstractFloat}
tree_p_from_flow(tree:: TreeLikeHS{FT}, flow::FT) where {FT<:AbstractFloat}

Calculate the leaf xylem end pressure, given

tree GrassLikeHS or PalmLikeHS or TreeLikeHS type hydraulics system

flow Given flow rate

Accordingly, there is a function tree_e_crit to calculate the critical flow rate for the whole plant. Be aware that function tree_p_from_flow and tree_e_crit only applies to the case of only one canopy layer (or big-leaf model). As to the case of multiple canopy layer, more functions are pending.

PlantHydraulics.tree_e_crit — Function

tree_e_crit(tree::PalmLikeHS{FT}, ini::FT) where {FT<:AbstractFloat}

Calculate the critical flow rate for the tree per LA, given

tree GrassLikeHS or PalmLikeHS or TreeLikeHS type hydraulics system

ini Initial guess of the critical flow rate